Move album controller functions to album plugin so they can be reused
